fix:gnomad link for x chromosome variants #429
Problem
Switch 23 -> X in gnomad link, as 23 does not work. The open targets link and UMich UK Biobank also have the same issue.

see : #429

Solution

*Gnomad*
Added a helper (23toX) for handlebar to format chromosomes.  Prefixing a variable with 23toX will convert 23 to X.
see : https://handlebarsjs.com/guide/builtin-helpers.html

*Open targets*
Currently, the URL below works
https://genetics.opentargets.org/variant/X_153635030_C_T
but this does not
https://genetics.opentargets.org/variant/23_153635030_C_T

*UMich UK Biobank*
Add a helper to test for X chromosome and use that to decide whether to render. Neither
http://pheweb.sph.umich.edu/SAIGE-UKB/variant/X-152900485-C-T
nor
http://pheweb.sph.umich.edu/SAIGE-UKB/variant/23-152900485-C-T
are work going to the pheontype page it also seems missing
https://pheweb.org/UKB-SAIGE/pheno/250.2