fix:selection on colocalization tables
The select table IDs changed after upgrading the Reacttable library, and the logic to determine the currently selected row had to be updated.

The Clinvar response changed, and the code had to be updated to handle this.

The clinvar response changed and the code had
to be updated to handle this.